location-importer: Speed up exporting the database
This optimisation returns the same result, but in about
one fifteenth of the time than the previous version of
the query.

This can also be parallelised by PostgreSQL so that more
CPU cores will decrease the time it takes to generate
the network list.

On my system this went down from 75 to 5 minutes.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>