debian: Add all temporary files to Gitignore
New packages have been added since the inception of the .gitignore and
as a result during build we see directories such as location-importer/
and files such as location-importer.debhelper.log.

This commit ensures all temporary subdirectories, and additional
generic build artifact files, are ignored by Git.

The subdirectory exceptions to this rule are:

- d/patches/ which may be used by Quilt
  considering the source format is '3.0 (quilt)',

- d/source/ for the format file,

- d/tests/ which may be used by autopkgtest
  to specify what test suites exist for the source.
  See: https://salsa.debian.org/ci-team/autopkgtest/-/raw/debian/5.16/doc/README.package-tests.rst

Signed-off-by: Valters Jansons <valter.jansons@gmail.com>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>