debian: Drop unintended files from location-python
_location.la gets built and installed to site-packages/, however
an .la file is not expected to reside in the Python root. Additionally,
the dependency library listed does not have its respective .la file
installed. Further complicating the situation, dh-python moves the
site-packages/ files to dist-packages/ silently which then results in
a broken libdir left behind in the .la file.

The only reason the file is there is that it gets built inside the
source directory, which gets copied entirely to location-python package
as-is. Considering the situation, this commit ensures the .la files is
not packaged by deleting it from the package files subdirectory.

location-importer package pulls in two Python (.py) files from the
source directory. These files should not be included in the
location-python package as a result.

Resolves: lintian:
  unknown-file-in-python-module-directory,
  incorrect-libdir-in-la-file,
  non-empty-dependency_libs-in-la-file,
  binaries-have-file-conflict

Signed-off-by: Valters Jansons <valter.jansons@gmail.com>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>