export: Fix filtering logic
It is possible to filter for what kind of network should be exported.

This worked well when the filter list only contained country codes, or
when it only contained ASNs. If there was a mix, only networks that
match both (i.e. virtually nothing) matched.

This patch fixes that we will use for either of them.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>