tests: Ensure loc_database_lookup finds a match
Commit 9e72b8a modified the lookup function to return 0 even if no
matches are found. As result, the test case could pass even if the
database lookup did not find any matches.

This commit restores the original intent of the test case, by adding
a failure condition when `network1` is NULL. This is similar to how
the subsequent test case was modified for the inverse condition.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>